+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * (breaking) IEP 03: mathematical and CSV syntax for Enumerations
 * (breaking) IEP 03: require `()` after constants and propositions
 * (breaking) IEP 03: use `in`, `∈` in all quantifications: `!x in Color`
+* (breaking) rename ``Symbols` into `Symbol` for consistency
 * More verbose error messages + line and col numbers (issue #64)
 * support for 'not in' operator
 * IEP 09 Unicode: allow 𝔹, ℤ, ℝ symbols for types, and '←' in definitions